Fixed: wflow-core/wflow-directory - Fixed possible NPE. @7.0-SNAPSHOT

+++ wflow-directory/src/main/java/org/joget/directory/dao/EmploymentDaoImpl.java
@@ -312,7 +312,7 @@ public class EmploymentDaoImpl extends AbstractSpringDao implements EmploymentDa
                     }
                 }
                 
-                if (ue.getEmploymentReportTo() != null) {
+                if (ue != null && ue.getEmploymentReportTo() != null) {
                     //delete current report to
                     employmentReportToDao.deleteEmploymentReportTo(ue.getEmploymentReportTo().getId());
                 }
